+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.0.2
+
+- Used deterministic deployments in migrations.
+- Updated dependency version `@animoca/ethereum-contracts@0.1.0`.
+- Added migrations for `ForwarderRegistryContextFacet` and `ERC721`-related facets.
+- Fixed bug in `skipIfFileExists`.
+
 ## 0.0.1
 
 - Initial release.